services/service/djangoapi/requirements.txt
Django>=4.2.9,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-gis>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.2,<5.0